+++ test/e2e/formbuilder/formbuilder.spec.js
@@ -931,6 +931,64 @@ describe('GET /', function () {
       expect(fb.updateFhir.isEnabled()).toBeFalsy();
       fb.dismissMenu();
     });
+
+    // Assume there is a resource on the server
+    it('should do next/previous page', function () {
+
+      fb.importMenu.click();
+      fb.showFhirResources.click();
+      expect(fb.dialog.isDisplayed()).toBeTruthy();
+      fb.fhirServerPulldownSelect(uhnServerName).click();
+
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeFalsy();
+      fb.nextButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.nextButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.nextButton.click();
+      expect(fb.nextButton.isEnabled()).toBeFalsy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+
+      fb.prevButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.prevButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.prevButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeFalsy();
+
+      fb.fhirServerPulldownSelect('UHN HAPI Server - STU3').click();
+
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeFalsy();
+      fb.nextButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.nextButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.nextButton.click();
+      expect(fb.nextButton.isEnabled()).toBeFalsy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+
+      fb.prevButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.prevButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeTruthy();
+      fb.prevButton.click();
+      expect(fb.nextButton.isEnabled()).toBeTruthy();
+      expect(fb.prevButton.isEnabled()).toBeFalsy();
+
+
+      fb.closeDialog(); // fhir results
+    });
   });
 
 });